feat(specialists): unify catalog and favorites into SpecialistFeed component
Add SpecialistFeed with mode='all'|'favorites'. Both /specialists and
/saved-specialists screens now delegate to this single component.
Closes #1588.
